requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Werkzeug==2.3.7
pandas==2.0.3
openpyxl==3.1.2
fpdf==1.7.2
gunicorn==21.2.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9